modify coupon BUG

+++ app/src/main/java/com/ipd/jianghuzuche/bean/CarReturnDetailsBean.java
@@ -409,8 +409,8 @@ public class CarReturnDetailsBean {
             private int rentDuration;
             private Object lateMoney;
             private int overdueMoney;
-            private int refundMoney;
-            private int defaultMoney;
+            private double refundMoney;
+            private double defaultMoney;
 
             public int getCostId() {
                 return costId;
@@ -516,19 +516,19 @@ public class CarReturnDetailsBean {
                 this.overdueMoney = overdueMoney;
             }
 
-            public int getRefundMoney() {
+            public double getRefundMoney() {
                 return refundMoney;
             }
 
-            public void setRefundMoney(int refundMoney) {
+            public void setRefundMoney(double refundMoney) {
                 this.refundMoney = refundMoney;
             }
 
-            public int getDefaultMoney() {
+            public double getDefaultMoney() {
                 return defaultMoney;
             }
 
-            public void setDefaultMoney(int defaultMoney) {
+            public void setDefaultMoney(double defaultMoney) {
                 this.defaultMoney = defaultMoney;
             }
         }
